feat: Add multivalency analysis and visualization
- Implemented multivalency analysis in `multivalency.py` using the Ule lab GeRMs R package.
- Created a new pipeline in `pipeline.py` to orchestrate loading, processing, and plotting of RNA maps.
- Added shared plotting helpers in `plot_helpers.py` for legend formatting and enrichment markers.
- Developed various plotting functions in `plots.py` for exon lengths, heatmaps, and RNA maps.
- Introduced preprocessing functions in `preprocessing.py` for subsetting, splice-site BED creation, and coverage smoothing.
- Enhanced logging throughout the pipeline for better traceability and debugging.
