+++ rnamaps/preprocessing.py
@@ -1,0 +1,109 @@
+"""Shared preprocessing: subsetting, splice-site BED creation, smoothing."""
+
+import logging
+
+import numpy as np
+import pandas as pd
+
+
+def apply_subsetting(df_rmats, no_constitutive):
+    """
+    Subset control and constitutive exons to match the largest regulated
+    category count. Returns (subsetted df, original_counts dict).
+    """
+    category_counts = df_rmats['category'].value_counts()
+    original_counts = {cat: count for cat, count in category_counts.items()}
+
+    target_count = 0
+    if 'enhanced' in category_counts and 'silenced' in category_counts:
+        target_count = max(category_counts['enhanced'], category_counts['silenced'])
+    elif 'enhanced' in category_counts:
+        target_count = category_counts['enhanced']
+    elif 'silenced' in category_counts:
+        target_count = category_counts['silenced']
+
+    # Subset control
+    if 'control' in category_counts and category_counts['control'] > target_count > 0:
+        control_indices = df_rmats[df_rmats['category'] == 'control'].index
+        control_keep = np.random.choice(control_indices, target_count, replace=False)
+        drop_mask = df_rmats.index.isin(control_indices) & ~df_rmats.index.isin(control_keep)
+        df_rmats = df_rmats[~drop_mask]
+        logging.info(f"Subsetted control exons from {category_counts['control']} to {target_count}")
+
+    # Subset constitutive
+    if (not no_constitutive
+            and 'constituitive' in category_counts
+            and category_counts['constituitive'] > target_count > 0):
+        const_indices = df_rmats[df_rmats['category'] == 'constituitive'].index
+        const_keep = np.random.choice(const_indices, target_count, replace=False)
+        drop_mask = df_rmats.index.isin(const_indices) & ~df_rmats.index.isin(const_keep)
+        df_rmats = df_rmats[~drop_mask]
+        logging.info(f"Subsetted constitutive exons from "
+                     f"{category_counts['constituitive']} to {target_count}")
+
+    return df_rmats, original_counts
+
+
+def get_ss_bed(df, pos_col, neg_col):
+    """
+    Create BED file for splice sites (handles strand orientation).
+
+    pos_col: column to use for + strand (one end of the exon)
+    neg_col: column to use for - strand (other end of same exon)
+
+    Both columns should reference the SAME exon. The upstream/downstream
+    strand correction happens at load time, not here.
+    """
+    df = df.copy()
+
+    df['exon_id'] = (
+        df['category'] + "_" +
+        df['chr'].astype(str) + ":" +
+        df['exonStart_0base'].astype(str) + "-" +
+        df['exonEnd'].astype(str) + ";" +
+        df['strand'].astype(str)
+    )
+
+    ss_pos = df.loc[df['strand'] == "+",
+                    ['chr', pos_col, pos_col, 'exon_id', 'FDR', 'strand']]
+    ss_pos.columns = ['chr', 'start', 'end', 'name', 'score', 'strand']
+    ss_pos.start = ss_pos.start.transform(lambda x: x - 1)
+
+    ss_n = df.loc[df['strand'] == "-",
+                  ['chr', neg_col, neg_col, 'exon_id', 'FDR', 'strand']]
+    ss_n.columns = ['chr', 'start', 'end', 'name', 'score', 'strand']
+    ss_n.end = ss_n.end.transform(lambda x: x + 1)
+
+    ss = pd.concat([ss_pos, ss_n])
+
+    return ss
+
+
+def smooth_coverage(df, window_size=10, std=2):
+    """Smooth coverage data using rolling Gaussian window."""
+    result = df.copy()
+    groups = []
+
+    for (exon_id, label), group_df in result.groupby(['exon_id', 'label']):
+        if len(group_df) < 3:
+            groups.append(group_df)
+            continue
+
+        group_sorted = group_df.sort_values('position')
+
+        if len(group_sorted) >= window_size:
+            values = group_sorted['coverage'].values
+            s = pd.Series(values)
+
+            smoothed = s.rolling(
+                window=window_size,
+                center=True,
+                win_type='gaussian'
+            ).mean(std=std)
+
+            smoothed = smoothed.fillna(s)
+            group_sorted['coverage'] = smoothed.values
+
+        groups.append(group_sorted)
+
+    return pd.concat(groups)
